requirements.txt
openai>=1.0.0
python-dotenv==1.0.0
pymongo[srv]==4.6.1
bcrypt==4.1.2
certifi>=2023.7.22
flask==3.0.0